Drop redundant bulletin tables, move controller from bulletin package to messages
