+++ service/src/main/kotlin/fi/espoo/evaka/messaging/message/ChildRecipientsController.kt
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
-package fi.espoo.evaka.messaging.bulletin
+package fi.espoo.evaka.messaging.message
 
 import fi.espoo.evaka.Audit
 import fi.espoo.evaka.shared.auth.AccessControlList
