+++ service/src/main/resources/db/migration/V111__drop_bulletin_tables.sql
@@ -1,0 +1,3 @@
+DROP TABLE bulletin_receiver;
+DROP TABLE bulletin_instance;
+DROP TABLE bulletin;
