+++ service/src/main/resources/migrations.txt
@@ -108,3 +108,4 @@ V107__scheduled_tasks.sql
 V108__person_freetext_search.sql
 V109__fee_thresholds_default_id.sql
 V110__varda_service_need.sql
+V111__drop_bulletin_tables.sql
